requirements.txt
matplotlib
tqdm
python-dotenv
pickle

# For deep learning
numpy
scipy
pandas
scikit-learn
pytorch-lightning>=2

# For database
mysqlclient # sudo apt install libmysqlclient-dev
sqlmodel
sqlalchemy
aiomysql

# For api server
fastapi
uvicorn[standard]
requests

# For logging
wandb
tensorboard
tensorboardX

# For graph
# graphtools # Installation instruction https://git.skewed.de/count0/graph-tool/-/wikis/installation-instructions#windows

# networkx
# igraph
# RAPIDS
# cudf